% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan-meiller.R
\name{find_coxph_cutoff}
\alias{find_coxph_cutoff}
\title{Find the cutoff for a given column in a data frame}
\usage{
find_coxph_cutoff(
  df,
  column,
  time,
  event,
  set_cutoff = NULL,
  covariates = NULL,
  interquartile = T
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{column}{The column name in the data frame that contains the grouping variable.}

\item{time}{The column name in the data frame that contains the time variable.}

\item{event}{The column name in the data frame that contains the event variable.}

\item{set_cutoff}{A numeric value specifying the cutoff to use. If NULL, the function will find the cutoff that maximizes the log-rank test p-value.}

\item{covariates}{A vector of column names in the data frame that contains the covariates.}

\item{interquartile}{Logical indicating whether to use the interquartile range to set the cutoff.}
}
\value{
A numeric value specifying the cutoff.
}
\description{
This function finds the cutoff for a given column in a data frame that maximizes the difference in survival between two groups.
It fits a Cox proportional hazards model and finds the cutoff that maximizes the log-rank test p-value.
}
