% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{make_multivariate_table}
\alias{make_multivariate_table}
\title{Make a multivariate table}
\usage{
make_multivariate_table(df, variables, covariates, time, event)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{variables}{A vector of column names in the data frame that contains the variables to include in the table.}

\item{covariates}{A vector of column names in the data frame that contains the covariates.}

\item{time}{The column name in the data frame that contains the time variable.}

\item{event}{The column name in the data frame that contains the event variable.}
}
\value{
A data frame containing the multivariate table.
}
\description{
This function makes a multivariate table for a given column in a data frame.
It fits a survival model and creates a table with the hazard ratios and p-values.
}
