% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph.R
\name{plot_forest_gene}
\alias{plot_forest_gene}
\title{Plot a forest plot for genes}
\usage{
plot_forest_gene(
  table,
  title = "",
  analysis = "multivariate",
  clip = c(0.3, 4)
)
}
\arguments{
\item{table}{A data frame containing the table.}

\item{title}{The title of the plot.}

\item{analysis}{The type of analysis to perform. Can be "multivariate" or "univariate".}

\item{clip}{The range of the x-axis.}
}
\value{
A ggplot object.
}
\description{
This function plots a forest plot for a given table.
}
