% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan-meiller.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot survival curves}
\usage{
plot_survival(
  df,
  column,
  time,
  event,
  title = "",
  plot = T,
  pval.coord = c(0.1, 0.1),
  legend.position = c(0.95, 0.425),
  ...
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{column}{The column name in the data frame that contains the grouping variable.}

\item{time}{The column name in the data frame that contains the time variable.}

\item{event}{The column name in the data frame that contains the event variable.}

\item{title}{The title of the plot.}

\item{plot}{Logical indicating whether to plot the survival curves.}

\item{pval.coord}{The coordinates of the p-value text.}

\item{legend.position}{The position of the legend.}

\item{...}{Additional arguments to pass to the ggsurvplot function.}
}
\value{
A ggplot object.
}
\description{
This function plots survival curves for a given column in a data frame.
It fits a survival model and creates a Kaplan-Meier plot with a risk table.
}
